+++ docs/changes.rst
@@ -20,6 +20,7 @@ New features
 - Add ``load_key`` and ``dump_key`` parameters to :class:`fields.Field` to control the field names
   during serialization/deserialization of data.
 - Add :attr:`ValidationError.state` to allow storing state to an error that can later be accessed.
+- Add ``include`` and ``exclude`` parameters to :meth:`Schema.dump`.
 
 Improvements
 ~~~~~~~~~~~~